Integration testing on book controller and update frontend

+++ backend/src/controllers/book.controller.ts
@@ -3,6 +3,7 @@ import { Request, Response, NextFunction, RequestHandler } from 'express';
 import prisma from '../../prisma/lib/prisma';
 import HttpException from '../config/exceptions/HttpException';
 import { IBook } from '../config/types';
+import { validEmail } from '@/utils/validation';
 
 const getAllBooks: RequestHandler = async (req: Request, res: Response, next: NextFunction) => {
   try {
@@ -28,9 +29,25 @@ const getBook: RequestHandler<{ id: string }> = async (req: Request, res: Respon
 };
 
 const createBook: RequestHandler = async (req: Request, res: Response, next: NextFunction) => {
-  const { title, author, type, year, feedBack, publisher, userMail } = req.body;
-
   try {
+    const { title, author, type, year, feedBack, publisher, userMail } = req.body;
+    const existingUser = await prisma.user.findUnique({ where: { email: userMail } });
+
+    if (!validEmail(userMail))
+      next(new HttpException(400, 'Invalid email.'));
+
+    if (!existingUser)
+      next(new HttpException(400, "User doesn't exist."));
+
+    if (!title || !author)
+      next(new HttpException(400, 'Missing data'));
+
+    if (feedBack && !Number(feedBack) || year && !Number(year))
+      next(new HttpException(400, 'Invalid number'));
+
+    if (feedBack < 0 || feedBack > 5)
+      next(new HttpException(400, 'Feedback invalid'));
+
     const newBook = await prisma.book.create({
       data: {
         title,
@@ -60,9 +77,28 @@ const createBook: RequestHandler = async (req: Request, res: Response, next: Nex
 };
 
 const updateBook: RequestHandler = async (req: Request, res: Response, next: NextFunction) => {
-  const { title, author, type, year, publisher, isRead, feedBack } = req.body;
-  const { id } = req.params;
   try {
+    const { title, author, type, year, publisher, isRead, feedBack, userId } = req.body;
+    const { id } = req.params;
+
+    if (!title || !author)
+        next(new HttpException(400, 'Missing data'));
+
+    if (feedBack && !Number(feedBack) || year && !Number(year))
+      next(new HttpException(400, 'Invalid number'));
+
+    if (feedBack < 0 || feedBack > 5)
+      next(new HttpException(400, 'Feedback invalid'));
+
+    const bookToUpdate = await prisma.book.findUnique({ where: { id } });
+    const currentUser = await prisma.user.findUnique({ where: { id: userId } });
+    
+    if (!bookToUpdate)
+      next(new HttpException(404, 'Book not found'));
+
+    if (bookToUpdate!.userId !== userId && currentUser?.role !== "ADMIN")
+      next(new HttpException(403, "User ID does not match book user ID"));
+
     const updatedBook = await prisma.book.update({
       where: { id },
       data: {
@@ -92,14 +128,20 @@ const updateBook: RequestHandler = async (req: Request, res: Response, next: Nex
   }
 };
 
-const deleteBook: RequestHandler<{ id: string }> = async (req: Request, res: Response, next: NextFunction) => {
+const deleteBook: RequestHandler = async (req: Request, res: Response, next: NextFunction) => {
   const { id } = req.params;
+
+  const bookToDelete = await prisma.book.findUnique({ where: { id } });
+
+  if (!bookToDelete)
+     next(new HttpException(404, 'Book not found'));
+
   try {
     const deletedBook: IBook = await prisma.book.delete({
-      where: { id: id },
+      where: { id },
     });
 
-    res.json(deletedBook);
+    res.status(204).json(deletedBook);
   } catch (error) {
     next(new HttpException(500, 'Something went wrong'));
   }
